Fish now are removed when intersect with player

+++ PajamaPenguins_Shared/SSKLibrary/SSKNodes/SSKScoreNode.m
@@ -58,7 +58,7 @@
 
 #pragma mark - Node's text
 - (void)updateText {
-    self.text = [NSString stringWithFormat:@"%lu",self.count];
+    self.text = [NSString stringWithFormat:@"%lu",(long)self.count];
 }
 
 @end
